`Cart.fromItems` now returns a cart type that more precisely encodes success/failure
The type `CartContentItem` has been refined to either `SuccessItem` or `FailedItem`.
A cart is either of type `InterimCart` which may have `FailedItem`s, or a `SuccessCart` which only
contains `SuccessItem`s.

`Cart.fromItems` validates the end result and returns either an `InterimCart` (if any failures happened)
or a `SuccessCart`. In the success case, the type does contain any error data
(as before, with `CartContentItems` that could in theory contain failed items)